Add first CircleCI config file.
Basic CI pipeline with code linting, format checking and unit testing.
An HTML code coverage report is generated as an artifact.

To be noted: the test coverage does not cover untested files. An open issue exists in the flutter repo (see
https://github.com/flutter/flutter/issues/27997). Left as is for now.

To be noted: the (private) firebase configuration files for cause commune and Thinkerview are missing in the CI.
Disabled related lint rules for the linter to run successfully in CI.
We could add these files as a private git submodule dependency (?).